refactor: move default values to class library
used class library to satisfy requirements. also removed unused library
