+++ DefaultValuesLibrary/DefaultValuesLibrary.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{20D8A0EA-D1B0-4AE8-8C9B-CFB96FC48B21}</ProjectGuid>
+    <ProjectGuid>{4DDAD5C8-439C-4DA5-9929-BD0EAD629C10}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>SearchSortUpdateLibrary</RootNamespace>
-    <AssemblyName>SearchSortUpdateLibrary</AssemblyName>
+    <RootNamespace>DefaultValuesLibrary</RootNamespace>
+    <AssemblyName>DefaultValuesLibrary</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -41,7 +41,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="SearchSortUpdate.cs" />
+    <Compile Include="DefaultStateValuesLibrary.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
